Test reading interestRate() within a batch
Within a batch, after performing an operation that affects the interest rate (ie, does a deposit/borrow/etc), calling interestRate() should immediately return the updated value, not the original value. This updated value can be compared against the value that interestRate() returns after the batch completes — they should be the same.

This will exercise the computeInterestRateView flow through to the IRM.